move python code to its own directory; split api from core blockchain class. use python 3.9
