+++ README.md
@@ -6,6 +6,8 @@ This is the source code for my post on [Building a Blockchain](https://medium.co
 
 ## Installation
 
+### Python
+
 1. Make sure [Python 3.6+](https://www.python.org/downloads/) is installed. 
 2. Install [pipenv](https://github.com/kennethreitz/pipenv). 
 
@@ -18,11 +20,11 @@ $ pipenv install
 ``` 
 
 4. Run the server:
-    * `$ pipenv run python blockchain.py` 
-    * `$ pipenv run python blockchain.py -p 5001`
-    * `$ pipenv run python blockchain.py --port 5002`
+    * `$ pipenv run python api.py` 
+    * `$ pipenv run python api.py -p 5001`
+    * `$ pipenv run python api.py --port 5002`
     
-## Docker
+#### Docker
 
 Another option for running this blockchain program is to use Docker.  Follow the instructions below to create a local Docker container:
 
@@ -47,17 +49,13 @@ $ docker run --rm -p 82:5000 blockchain
 $ docker run --rm -p 83:5000 blockchain
 ```
 
-## Installation (C# Implementation)
+### C#
 
 1. Install a free copy of Visual Studio IDE (Community Edition):
 https://www.visualstudio.com/vs/
-
 2. Once installed, open the solution file (BlockChain.sln) using the File > Open > Project/Solution menu options within Visual Studio.
-
 3. From within the "Solution Explorer", right click the BlockChain.Console project and select the "Set As Startup Project" option.
-
 4. Click the "Start" button, or hit F5 to run. The program executes in a console window, and is controlled via HTTP with the same commands as the Python version.
-
 
 ## Contributing
