+++ python/.travis.yml
@@ -1,7 +1,7 @@
 language: python
 
 python:
-  - 3.6
+  - 3.9
   - nightly
 
 install:
