+++ python/Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.6-alpine
+FROM python:3.9-alpine
 
 WORKDIR /app
 
@@ -9,7 +9,8 @@ RUN cd /app && \
 
 # Add actual source code.
 ADD blockchain.py /app
+ADD api.py /app
 
 EXPOSE 5000
 
-CMD ["python", "blockchain.py", "--port", "5000"]
+CMD ["python", "api.py", "--port", "5000"]
