--- requirements.txt
@@ -1,2 +1,0 @@
-flask==0.12.2
-requests==2.18.4
